Organize QR module
