+++ collect_app/src/main/java/org/odk/collect/android/application/Collect.java
@@ -24,7 +24,7 @@ import androidx.annotation.NonNull;
 
 import org.jetbrains.annotations.NotNull;
 import org.odk.collect.android.dynamicpreload.ExternalDataManager;
-import org.odk.collect.qrcode.MlKitBarcodeScannerViewFactory;
+import org.odk.collect.qrcode.mlkit.MlKitBarcodeScannerViewFactory;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.injection.config.AppDependencyComponent;
 import org.odk.collect.android.injection.config.CollectDrawDependencyModule;
