+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -52,7 +52,7 @@ import org.odk.collect.android.formmanagement.FormsDataService;
 import org.odk.collect.android.formmanagement.OpenRosaClientProvider;
 import org.odk.collect.android.formmanagement.ServerFormsDetailsFetcher;
 import org.odk.collect.qrcode.BarcodeScannerViewContainer;
-import org.odk.collect.qrcode.PlayServicesMlKitBarcodeScannerViewFactory;
+import org.odk.collect.qrcode.mlkit.PlayServicesMlKitBarcodeScannerViewFactory;
 import org.odk.collect.android.geo.MapConfiguratorProvider;
 import org.odk.collect.android.geo.MapFragmentFactoryImpl;
 import org.odk.collect.android.instancemanagement.InstancesDataService;
@@ -127,9 +127,9 @@ import org.odk.collect.projects.ProjectCreator;
 import org.odk.collect.projects.ProjectsRepository;
 import org.odk.collect.projects.SettingsConnectionMatcher;
 import org.odk.collect.projects.SharedPreferencesProjectsRepository;
-import org.odk.collect.qrcode.QRCodeCreatorImpl;
-import org.odk.collect.qrcode.QRCodeDecoder;
-import org.odk.collect.qrcode.QRCodeDecoderImpl;
+import org.odk.collect.qrcode.zxing.QRCodeCreatorImpl;
+import org.odk.collect.qrcode.zxing.QRCodeDecoder;
+import org.odk.collect.qrcode.zxing.QRCodeDecoderImpl;
 import org.odk.collect.settings.ODKAppSettingsImporter;
 import org.odk.collect.settings.ODKAppSettingsMigrator;
 import org.odk.collect.settings.SettingsProvider;
