+++ qr-code/src/main/java/org/odk/collect/qrcode/mlkit/PlayServicesMlKitBarcodeScannerViewFactory.kt
@@ -1,7 +1,10 @@
-package org.odk.collect.qrcode
+package org.odk.collect.qrcode.mlkit
 
 import android.app.Activity
 import androidx.lifecycle.LifecycleOwner
+import org.odk.collect.qrcode.BarcodeScannerView
+import org.odk.collect.qrcode.BarcodeScannerViewContainer
+import org.odk.collect.qrcode.zxing.ZxingBarcodeScannerViewFactory
 
 class PlayServicesMlKitBarcodeScannerViewFactory : BarcodeScannerViewContainer.Factory {
